Disable building with liburcu on powerpc
gdnsd currently FTBFS on powerpc, ppc64 and ppc64el because of a test
suite failure in test /015zruntime/042multi.t. The issue is reproducible
on powerpc and ppc64el porterboxes running as VMs on POWER8 (pizzetti
and plummer), but not on a POWER7 box (partch).

Further investigation showed that the reload thread ("gdnsd-zones") is
getting stuck in synchronize_rcu_qsbr() which in turn waiting for a
futex (running the futex() syscall). The issue was reproduced with both
sid's liburcu (0.9.1-3) and current liburcu master (4d0d7cb), both
running with a 3.16.7-ckt20-1+deb8u4 kernel (so no membarrier()
support). Hand-patching liburcu to not define CONFIG_RCU_HAVE_FUTEX did
fix the issue, but I stopped investigating there as I'm in over my head.

It's worth noting that this is not a gdnsd regression per se, as a
rebuild of 2.2.0-1 on the same boxes fails. However, that version was
previously built successfully on a POWER8 VM among others, while running
the same liburcu version. This test suite failure is also not happening
on any other architecture but those three powerpc ones (powerpcspe is
missing B-Ds, so it's an unknown so far). It's thus likely that this is
a kernel/qemu/toolchain issue but one that is probably hard to pinpoint
at this point.

For now, pass --without-urcu to ./configure on those architectures, as
this is an optimization we can live without. This may be reverted at
some point later when those issues have been fixed.